Consolidate categories & tags into single template
