+++ wp-templates/index.js
@@ -1,15 +1,15 @@
-import category from './category';
-import tag from './tag';
+// import category from './category';
+// import tag from './tag';
 import frontPage from './front-page';
 import page from './page';
 import single from './single';
 import project from './project';
+import archive from './archive';
 
 export default {
-  category,
-  tag,
   'front-page': frontPage,
   page,
   single,
   project,
+  archive,
 };
